backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1

# AI / RAG Stack
langchain==0.3.13
langchain-text-splitters==0.3.4
langchain-groq==0.2.4
langchain-google-genai==2.0.8
langchain-community==0.3.13
chromadb==0.5.23
sentence-transformers==3.3.1
